feat: Combine preprocessing and crawling codes
